Temperature data read from Commander CORE.
AIO Pump speed read from Commander CORE.

Fan still read only for now.

Possible Commander CORE XT support - requires willing victim :)

+++ FanControl.CommanderPro/Pro/CommanderPro.cs
@@ -291,9 +291,9 @@ namespace FanControl.CommanderPro.Pro
             return result;
         }
 
-        public Int32 GetTemperature(Int32 channel)
+        public Single GetTemperature(Int32 channel)
         {
-            Int32 result = 0;
+            Single result = 0;
 
             if (IsConnected)
             {
